### 2018-08-18 #26 @whedon
[PRE REVIEW]: Mikrokosmos: an educational lambda calculus interpreter
**Submitting author:** @mroman42 (<a href="http://orcid.org/0000-0003-3158-1226">Mario Román</a>)
**Repository:** <a href="https://github.com/mroman42/mikrokosmos" target ="_blank">https://github.com/mroman42/mikrokosmos</a>
**Version:** v0.8.0
