### 2018-08-21 #26 @RobertTalbert
@mroman42 Everything looks to be in order with the submission. I'll be looking for reviewers for it presently. BTW this looks extremely cool -- I was teaching myself $\lambda$-calculus last summer and wished at the time for some way of checking my work. This would have been perfect!
