### 2018-08-21 #26 @RobertTalbert
Thanks @MattForshaw. I don't believe we have hard-and-fast time limitations on these reviews but obviously sooner is better. I would say within 30 days, but if you need more time I think that would be OK as long as it doesn't drag on. @labarba Correct me if I'm wrong on anything here.
