### 2018-08-21 #26 @RobertTalbert
@MattForshaw Here's the documentation on the review process to give you a sense of what's involved: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
