### 2018-08-21 #26 @labarba
@RobertTalbert Are you going to want a second reviewer here, or are happy with just one (plus your editorial review)?
