### 2018-08-21 #26 @RobertTalbert
@labarba I solicited a second reviewer (@ahmaurya) but haven't heard back yet. Do I need to assign both reviewers at the same time? It looked in the editorial guide like I could do one now and the other later.
