### 2018-08-21 #26 @labarba
Best to wait for a couple of days for a response from the potential second reviewer.

Then you add them here (`@whedon add @username as reviewer`) — then start the review (`@whedon start review`). That way, each reviewer gets their own checklist on the review issue.
