### 2018-08-27 #26 @RobertTalbert
Update: I sent an email to @ahmaurya to see if he's interested in reviewing. I'm giving it a day or two, and then if I haven't heard back I'll move on to another second reviewer.
