### 2018-08-27 #26 @ahmaurya
Hi. I am not very familiar with lambda-calculus. As a result, I need to decline the review request especially because there will be many more people far better equipped to perform this review.
