### 2018-08-28 #26 @RobertTalbert
I've contacted another person for the second reviewer slot. Again, I'll give it a day or two and then we'll decide what to do from here. If finding a person who feels qualified to look at lambda calculus is an issue, I may do the second review myself or just stick with one reviewer.
