### 2018-09-02 #29 @whedon
[REVIEW]: Mikrokosmos: an educational lambda calculus interpreter
**Submitting author:** @mroman42 (<a href="http://orcid.org/0000-0003-3158-1226">Mario Román</a>)
**Repository:** <a href="https://github.com/mroman42/mikrokosmos" target ="_blank">https://github.com/mroman42/mikrokosmos</a>
**Version:** v0.8.0

@aj2duncan,  @MattForshaw please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
