### 2018-09-02 #26 @RobertTalbert
close https://github.com/openjournals/jose-reviews/issues/26
