### 2018-09-02 #26 @RobertTalbert
Apparently I don't know how to close an issue. Help?
