### 2018-09-02 #29 @labarba
OK, @RobertTalbert — I manually added a second review checklist above. We're good to go.

👋  @aj2duncan, @MattForshaw — Thank you for agreeing to review for JOSE! This is where the actin happens. Each of you has a review checklist above. Feel free to ask questions here, and to open new issues on the submission's repository, if you feel the need.
