### 2018-09-02 mroman42/mikrokosmos#45 @aj2duncan
Jupyter Notebook kernel not working on Windows
The command line version installs and works just fine on my Windows machine but there seems to be an issue with the Jupyter kernel. Everything installed fine but when I run it the kernel dies and cannot restart. Looking at what is trying to be run it seems the line
```
mikro = pexpect.spawn('mikrokosmos')
```
is causing the problem. The next line is
```
AttributeError: module 'pexpect' has no attribute 'spawn'
```
A little digging led to [https://pexpect.readthedocs.io/en/stable/overview.html#pexpect-on-windows](https://pexpect.readthedocs.io/en/stable/overview.html#pexpect-on-windows) which states that `spawn` is not available on Windows.

I'm not sure what the best way to proceed here. I guess the easiest option would be to make it clear that it doesn't work on Windows (the command line version works very well so I don't really mind) but you may feel differently.

The other option is I'm missing something entirely. If that is the case, please just let me know.
