### 2018-09-02 #29 @aj2duncan
I've had some difficulty getting the Jupyter Notebook version working on Windows. Please see [mroman42/mikrokosmos#45](https://github.com/mroman42/mikrokosmos/issues/45) Anyone else have this problem?

Command line version installs and works just fine.
