### 2018-09-03 0f987cf
Makes it clear that the Jupyter kernel does not work on Windows
From #45.
