### 2018-09-04 mroman42/mikrokosmos#47 @MattForshaw
Need for a prominent statement showing where users should seek support.
Good afternoon,

I am currently running through the JOSE review process, and would to suggest you add a prominent statement showing where users should seek support with mikrokosmos. For example, you may prefer users to reach out to you using Github Issues, or another mechanism.

Best wishes,
Matt
