### 2018-09-04 #29 @MattForshaw
Suggesting the authors add a more prominent statement where users should seek support. https://github.com/mroman42/mikrokosmos/issues/47
