### 2018-09-04 8c0bc57
Adds statement showing where to seek support
Addresses #47
