### 2018-09-04 mroman42/mikrokosmos#47 @mroman42
Good afternoon,

Thank you very much. Commit 8c0bc57d5e8329e76dbcb75c656f1e228c6446e7 addresses this problem.  It adds a section to the user's guide discussing where to seek support and contributing and a similar statement to the README of the main repository.  Please let me know if you have any suggestion on the format of the statement or on where should it be placed to be more easily seen by new users; this is very useful to me.

Best wishes,
Mario
