### 2018-09-05 mroman42/mikrokosmos#45 @mroman42
Using the newly created `windows-support` [branch](https://github.com/mroman42/jupyter-mikrokosmos/tree/windows-support), the kernel seems to work correctly under Windows10. However, some Unicode characters are not displayed correctly. These problems seem not easy to solve right now, so I have added a short note to the [readme](https://github.com/mroman42/jupyter-mikrokosmos/blob/master/README.md) file explaining that Windows users will need to download that specific branch and that unicode output may not be correctly displayed.
