### 2018-09-05 mroman42/mikrokosmos#45 @aj2duncan
Thanks for sorting the install and including the much clearer instructions for users. It seems to work exactly as you describe. The only other thing I might suggest is to add some instructions for installing from a specific branch. I used
```
pip install git+https://github.com/mroman42/jupyter-mikrokosmos.git@windows-support
```
but I imagine there are alternatives. I'll leave it to you whether you close the issue or not - it does install and function but there are still the unicode problems. Thanks again.
