### 2018-09-09 7ba0bd7
How to to install from the Github repository
From #45.
