### 2018-09-09 mroman42/mikrokosmos#45 @mroman42
After checking that the 'windows-support' kernel also works with Linux, I have merged the two branches and the new version of the Juptyer kernel has been uploaded to [PyPi](https://pypi.org/project/imikrokosmos/#history) (mroman42/jupyter-mikrokosmos@280c2b3).  The kernel with Window support can be now directly downloaded using `pip install imikrokosmos` (or using `pip install imikrokosmos -U` if a previous version was already installed). I hope this will be easier for users than installing the kernel from a specific branch.

In 7ba0bd7419eaf63829c425e9e04aa8812c98de25, I add a comment to the user's guide describing how to install the kernel using the source code of a specific branch of the Github repo using your command. Thank you very much for this suggestion.

Finally, I have opened a specific issue (#49) for the problem with Unicode characters. I prefer to close this issue and keep the other one open until a solution is found. Please reopen this issue if there is any problem or suggestion that I have left unaddressed.
