### 2018-09-09 #29 @aj2duncan
@RobertTalbert I'm not entirely sure how it works from here but I'm pretty happy that @mroman42 has satisfied the only issue I had. I really like the project and think the documentation is very good. Let me know if you need anything else from me.
