### 2018-09-09 mroman42/mikrokosmos#45 @aj2duncan
Hi @mroman42 thanks very much for sorting this and I think you have dealt with everything.
