### 2018-09-09 #29 @labarba
OK.

👋 @RobertTalbert --- You can give the paper an editorial check now, suggest any improvements you'd like to see, and then let me know you're ready to accept.
