### 2018-09-18 #29 @RobertTalbert
@labarba I've had a look at the paper and have no other suggestions, so I think this one is good to go.
