### 2018-10-02 #29 @labarba
@mroman42 — could you tweak the paper as follows?

1. paragraph 3: "Following (Barendregt, 1984)" ... if the reference is part of the sentence, it should be formatted as: "Following Barendregt (1984)"

2. "as described in (De Bruijn, 192)" ... here, you can remove the "as described in" and simply put the reference at the end of the sentence. It will make it more readable, given the large distance you have between subject and verb ("works").

3. "standard algorithm from (Hindley & Seldin, 2008)" ... same as 1.

4. "see (Sorensen & Urzyczyn, 2006)" ... same.

5. "textbooks such as (Wadler, 2015) or (Girard, Taylor, & Lafont, 1989)" ... same

6. page 2: "installation and usage instructions ... are also described" ... change verb to "given" or "provided"?

7. "described in (Wadler, 2015)" ... "depicted in (Wadler, 2007)" ... "justified in (Danierlson, Hughes, Jansson, & Gibbons, 2006)" ... same as 1.

8. Find other examples in page 3.
