### 2018-10-03 #29 @mroman42
Thank you very much. Commit mroman42/mikrokosmos@a5e9c879cfa9b35fe99c786bc61fad3bf8775b56 addresses these corrections and changes some propositions to make the name of the author appear as part of the sentence. Please let me know if any other change is required.
