### 2018-10-23 #29 @labarba
@mroman42 — apologies for the delay (I just came back this weekend from a three-talk tour and I was slammed).

We're ready to publish your submission, once you deposit an archive on Zenodo (or a similar service: Figshare, Dryad, etc.). Once you get a DOI for the archive, give it to us here, and we'll publish your JOSE paper!
