+++ README.md
@@ -1,7 +1,9 @@
 # mikrokosmos
 
-![Version](https://img.shields.io/badge/version-0.8.0-blue.svg)
+![Version](https://img.shields.io/badge/version-0.8.1-blue.svg)
 ![Travis (.org)](https://img.shields.io/travis/mroman42/mikrokosmos.svg)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1470226.svg)](https://doi.org/10.5281/zenodo.1470226)
+
 
 **Mikrokosmos** is a λ-calculus interpreter, borrowing its name from the series of
 progressive piano études *[Mikrokosmos](https://www.youtube.com/watch?v=VEsMk3DAzWM)* written by *Bela Bartok*. 
