### 2018-10-24 #29 @mroman42
If I have done everything correctly (please let me know if something else is needed) the DOI is 10.5281/zenodo.1470226 (https://zenodo.org/record/1470226).

@labarba @RobertTalbert @aj2duncan @MattForshaw Thank you very much for all the feedback; it has been very useful for the project.
