### 2018-10-24 #29 @labarba
Almost ... the author list on the Zenodo archive does not match the author list of the paper. On Zenodo, we have "Mario Román; Pedro A. García-Sánchez" ... on the JOSE paper, just Mario Román. You may need to manually change the metadata in the Zenodo entry.
