### 2018-10-24 #29 @labarba
This paper is now published in JOSE !!!

@aj2duncan, @MattForshaw — Thank you for your review and being a part in this adventure in new publishing!
